[AC-7729] Raise assert_notified assertions to api testcase
