+++ web/impact/impact/tests/api_test_case.py
@@ -6,6 +6,7 @@ from oauth2_provider.models import get_application_model
 from rest_framework.test import APIClient
 from test_plus.test import TestCase
 
+from django.core import mail
 from django.conf import settings
 from django.contrib.auth.models import Group
 from django.urls import reverse
@@ -106,3 +107,23 @@ class APITestCase(TestCase):
             data['header'] == header,
             data['detail'] == notification
         ]), msg='Notification data was not as expected')
+
+    def assert_notified(self, user, message=""):
+        '''Assert that the user received a notification. 
+        If `message` is specified, assert that the message appears in one of 
+        the outgoing emails to this user
+        '''
+        emails = [email for email in mail.outbox if user.email in email.to]
+        self.assertGreater(len(emails), 0)
+        if message:    
+            self.assertTrue(any([message in email.body for email in emails]))
+
+
+    def assert_not_notified(self, user):
+        '''Assert that the specified user did not receive a notification.
+        '''
+        if mail.outbox:
+            self.assertNotIn(user.email, [email.to for email in mail.outbox],
+                             msg="Found an email sent to user")
+
+
